Reorganized files for future additions
