+++ glossary/glossary.php
@@ -1,0 +1,27 @@
+<?php
+/**
+ * WDS Site Documentation.
+ *
+ * @package           WebDevStudios\Documentation
+ * @author            WebDevStudios
+ * @copyright         2021 WebDevStudios
+ * @license           GPL-2.0-or-later
+ * @since             1.1.1
+ *
+ * @wordpress-plugin
+ * Plugin Name:       WDS Site Documentation
+ * Plugin URI:        https://github.com/webdevstudios/wds-site-documentation
+ * Description:       A plugin to host site documentation in an easily accessible place in the WordPress dashboard.
+ * Version:           1.1.1
+ * Requires at least: 5.2
+ * Requires PHP:      7.2
+ * Author:            WebDevStudios
+ * Author URI:        https://webdevstudios.com
+ * Text Domain:       wds-site-documentation
+ * License:           GPL v2 or later
+ * License URI:       http://www.gnu.org/licenses/gpl-2.0.txt
+ */
+
+namespace WebDevStudios\Documentation;
+
+use WP_Query;
